Improve the Developer experience by automatic code compiling
By adding a watch task for our client side code we can improve the
speed of the compiles as webpack caches results and only applies
updates to parts of the code that changes.

Using the forever module we can watch application files for changes
are restart the server, when they do.

And finally using concurrently we can run these two tasks at the same
 time exposing a `npm run dev` | `yarn dev` on the project. While we
 develop.