Add CSS capabilities to the application.
The goal is to be able to place the css next to each component that
uses it. Doing so means that the it's simpler for
developers to see what css belongs to which components. If a
component is removed is just as trivial to remove.

https://github.com/azaharakis/ui-training/tree/master/src/5.%20Encapsulating%20Components
explains more around the idea of encapsulation of features.

Why this is relevant here is because we need to ignore css on the server
as the files are referenced within the dependency tree of each
component. This will become clearer when a component is implemented with
css.

We also setup webpack to be able to compile the css. through the use
of 2 loaders. style-loader and css-loader

takes the CSS and returns the CSS with imports and url(...)
resolved via webpack's require functionality

takes CSS and actually inserts it into the page so that the styles are active on the page.
This is done via javascript, in a production scenario we'd want to
pre-compile into a css file, this can be done via, you guessed it
another loader. But this is outside the scope of this commit
 Doing so