Setting up an API call using Fetch
While the Fetch API is still an experimental technology, its
generally used in the industry as a light weight way to make API
call's it's also already supported by some browsers. For those
browsers that don't support it we are using a polyfill. Also as we
will eventually make API calls on the server it's a good idea to have
an API that is consistent across both client and server. Which is
another reason for this choice of polyfill.

We have chosen to implement the fetching of data from the API in a
lifecycle method of the App component. `componentDidMount` executes
immediately after the initial render on the client.

https://github.com/azaharakis/ui-training/tree/master/src/3.%20React/2.%20Managing%20State
Explains how the App component uses state to update the view of the
client once the response has returned.