Fetch API data on the server
One of the benefits of having a server is that we can set it up to
render a full page response. The problem with our last commit
was that we required the user to download a javascript package
to make a request to an API to display a list of properties.
Meaning
- We unnecessarily delay the rendering of meaningful content
- We make it harder for SEO Crawlers to access our content.

With a few adjustments we can change our implementation so that we
make the API call on the server. Populate the data within a variable as
part of the payload we send to the browser, and then use that value to
"hydrate" our application for the client.

Because the API request is async we need to execute it at part of the
application lifecycle that will wait before sending a response. Thus we
create a static method on the React class, which we can execute without
an instance from the middleware module. Once the data has returned we
then resolve the Promise and send a request to the browser.