Introduction of Redux
Two dependencies have been introduced:
redux - A library for managing application state
react-redux - A library that provides bindings to use redux with your
 react components

This redux implementation is to hold our apiData within a javascript
store, a fair amount of boilerplate has been added just to do this
and it's hard to see the value it's providing at the moment.
However if we want to add additional screens to the application,
with or without their own data requirements. Using redux starts to
become advantageous particularly in conjunction with react-router, as
 the <Provider /> component uses a feature of react called [context]
 (https://facebook.github.io/react/docs/context.html) that allows you
  to 'Connect' components to that data, as you can see from the newly
  renamed App -> Container component.