Added a Client | Server routing using react-router
- Creation of root level route (app)
- Two additional screens added to showcase routing
   - Results that has an API call to fetch data
   - Homepage which is a basic react component that links to results

`./src/client/index` and `./src/server/middleware` glues this together
and allows the application to serve requests in a way that's diagrammed
below:
 https://community.rea-group.com/servlet/JiveServlet/showImage/38-19968-5648757/universal+app.png

in `./src/server/index` we changed `get` to `use` and actually applied
middleware.js as middleware, we send all requests that aren't
captured elsewhere to this middleware to be processed by react-router.
If we don't understand it we have a catch all route `*` that returns
a <div>404</div>

`./src/shared/fetchData` is a module that is executed on both client
and
the server to execute the current route's API call if there is one.
It can do this because we assign a prop `fetchData` on the route
within `./src/shared/routes`

`./src/shared/createAppFromRedux` is removed as we need a different
implementation on the server to the client with react-router. On the
server the router needs no concept of state, it matches the current
request that comes in to a route makes an API call and then sends the
response to the browser. So we use RouterContext as recommended by
react-router https://github.com/reactjs/react-router-tutorial/tree/master/lessons/13-server-rendering

On the client we use the Router Component, similar to RouterContext
component but it maintains the state internally in the Router, things
 like browserHistory.

browserHistory is another interesting module and it is exposed by
react-router, it is a library that decorates the browser history API
https://github.com/ReactTraining/react-router/blob/master/docs/guides/Histories.md
We bind an event in `./src/client/index` to listen on history change
so we can dispatch events to the redux store to update the application.