Add Testing frameworks and tests to the App Container
- jest: a framework for running, and writing tests. In this case it
is allowing us to setup the Container.test file, it allows us to
write our tests using `describe` and `it` functions and it executes
and provides a report on the status of the tests.
- babel-jest: Can be seen as the equivalent of a loader for webpack
but is used in the conext of jest. When the jest command is run it
will use this plugin to compile the javascript code using the
settings defined in the .babelrc file.
- enzyme: Gives us handy features when it comes to working with react
components. In this case we are mounting our components using enzyme
in our test environment and accessing their values using methods
that enzyme exposes.
- react-addons-test-utils Is used under the hood by enzyme and is a
library that react provides. We don't need to use it directly.

We also added the tests to the test task and created a watch task
which will further enable TDD.

The jest entry in the package.json performs a simialr tole to what a
webpack file would do. In this case when any of the extensions within
the moduleNameMapper are found in the tests, they will be stubbed with
empty files. We won't be performing any unit tests on css or media items
such as images.