Simulate a slow API and add a Loader Component
When transitioning between asynchronous screen's we'll want to
provide feedback to the user that the application is indeed loading
the next page. As the API we have built is local, the data will
almost always be instantaneous, to more accurately see this
functionality in action, we put a timeout before resolving the fetch
response. Because Fetch uses a Promise like API we can just return a
promise within the Results Container, that delays the fetch.