Battle harden the server for erroneous requests
Out application was quite naive and wouldn't handle anything other
then a matched route. Using the redirect parameter that is passed in
through the match callback. We can use express to redirect users if
we specify a redirect route in our `./src/shared/routes.js` module.

If routeing state doesn't exist we can return a 404 back to the user
telling them we don't have a screen for the requested URL.

And finally if we have a run time error creating our screen we can
specify a catch on our promise, log the error and return a 500